Address review comments.

+++ SmartDeviceLink/SDLMassageCushionFirmness.m
@@ -9,7 +9,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation SDLMassageCushionFirmness
 
-- (instancetype)initWithMassageCushion:(SDLMassageCushion)cushion firmness:(UInt16)firmness {
+- (instancetype)initWithMassageCushion:(SDLMassageCushion)cushion firmness:(UInt8)firmness {
     self = [super init];
     if (!self) {
         return nil;
